MAINT add base class for voting and stacking
